+++ qclib/state_preparation/cvoqram.py
@@ -21,12 +21,10 @@ from qiskit.quantum_info import Operator
 from qclib.util import _compute_matrix_angles
 from qclib.gates.initialize_sparse import InitializeSparse
 from qclib.gates.mcg import Mcg
+from qclib.gates.ldmcsu import LdMcSpecialUnitary
 
 
 # pylint: disable=maybe-no-member
-
-
-QuantumCircuit.mcg = Mcg.mcg
 
 
 class CvoqramInitialize(InitializeSparse):
@@ -47,17 +45,17 @@ class CvoqramInitialize(InitializeSparse):
         default_with_aux = True
         if opt_params is None:
             self.with_aux = default_with_aux
-            self.use_linear_mcg = True
+            self.mcg_method = 'linear'
         else:
             if opt_params.get("with_aux") is None:
                 self.with_aux = default_with_aux
             else:
                 self.with_aux = opt_params.get("with_aux")
 
-            if opt_params.get("use_linear_mcg") is None:
-                self.use_linear_mcg = True
+            if opt_params.get("mcg_method") is None:
+                self.mcg_method = 'linear'
             else:
-                self.use_linear_mcg = opt_params.get("use_linear_mcg")
+                self.mcg_method = opt_params.get("mcg_method")
 
         if label is None:
             label = "CVOSP"
@@ -150,11 +148,14 @@ class CvoqramInitialize(InitializeSparse):
                 self._mcuvchain(circuit, theta, phi, lam)
             else:
                 gate = UGate(theta, phi, lam)
-                if self.use_linear_mcg:
-                    gate_op = Operator(gate).data
-                    circuit.mcg(gate_op, memory[control], [self.aux[0]])
-                else:
+                if self.mcg_method == 'qiskit':
                     circuit.append(gate.control(len(control)), memory[control] + [self.aux[0]])
+                elif self.mcg_method == 'barenco':
+                    gate_op = Operator(gate).data
+                    LdMcSpecialUnitary.ldmcsu(circuit, gate_op, memory[control], [self.aux[0]])
+                else:
+                    gate_op = Operator(gate).data
+                    Mcg.mcg(circuit, gate_op, memory[control], [self.aux[0]])
 
         self.norm = self.norm - np.absolute(np.power(feature, 2))
