+++ test/gates/test_majority.py
